+++ x/evm/keeper/statedb.go
@@ -654,7 +654,7 @@ func (k *Keeper) AddLog(log *ethtypes.Log) {
 
 	log.Index = uint(k.GetLogSizeTransient())
 	k.IncreaseLogSizeTransient()
-	k.SetLog(log)
+	k.AddLogTransient(log)
 
 	k.Logger(ctx).Debug(
 		"log added",